Cache condition regexes
