+++ tracer/src/Datadog.Trace/FeatureFlags/FeatureFlagsEvaluator.cs
@@ -268,9 +268,9 @@ namespace Datadog.Trace.FeatureFlags
             switch (condition.Operator)
             {
                 case ConditionOperator.MATCHES:
-                    return MatchesRegex(attributeValue, condition.Value);
+                    return condition.MatchesRegex(attributeValue);
                 case ConditionOperator.NOT_MATCHES:
-                    return !MatchesRegex(attributeValue, condition.Value);
+                    return !condition.MatchesRegex(attributeValue);
                 case ConditionOperator.ONE_OF:
                     return IsOneOf(attributeValue, condition.Value);
                 case ConditionOperator.NOT_ONE_OF:
@@ -285,32 +285,6 @@ namespace Datadog.Trace.FeatureFlags
                     return CompareNumber(attributeValue, condition.Value, (a, b) => a < b);
                 default:
                     throw new FormatException($"Unknown condition operator {condition.Operator.ToString()}");
-            }
-        }
-
-        private static bool MatchesRegex(object attributeValue, object? conditionValue)
-        {
-            if (conditionValue is null)
-            {
-                throw new FormatException("Condition value can not be null");
-            }
-
-            try
-            {
-                var pattern = conditionValue?.ToString() ?? string.Empty;
-                var regex = new Regex(pattern);
-                return regex.IsMatch(ToString(attributeValue));
-            }
-            catch
-            {
-                return false;
-            }
-
-            static string ToString(object attributeValue)
-            {
-                if (attributeValue is null) { return string.Empty; }
-                if (attributeValue is bool boolValue) { return boolValue ? "true" : "false"; }
-                return Convert.ToString(attributeValue, CultureInfo.InvariantCulture) ?? string.Empty;
             }
         }
